+++ runtime/src/snapshot_utils/archive_format.rs
@@ -4,39 +4,26 @@ use {
 };
 
 // SUPPORTED_ARCHIVE_COMPRESSION lists the compression types that can be
-// specified on the command line. "zstd" and "lz4" are valid whereas "gzip",
-// "bz2", "tar" and "none" have been deprecated. Thus, all newly created
-// snapshots will either use "zstd" or "lz4". By keeping the deprecated types
-// in the ArchiveFormat enum, pre-existing snapshot archives with the
-// deprecated compression types can still be read.
+// specified on the command line.
 pub const SUPPORTED_ARCHIVE_COMPRESSION: &[&str] = &["zstd", "lz4"];
 pub const DEFAULT_ARCHIVE_COMPRESSION: &str = "zstd";
 
-pub const TAR_BZIP2_EXTENSION: &str = "tar.bz2";
-pub const TAR_GZIP_EXTENSION: &str = "tar.gz";
 pub const TAR_ZSTD_EXTENSION: &str = "tar.zst";
 pub const TAR_LZ4_EXTENSION: &str = "tar.lz4";
-pub const TAR_EXTENSION: &str = "tar";
 
 /// The different archive formats used for snapshots
 #[derive(Copy, Clone, Debug, Eq, PartialEq, Display)]
 pub enum ArchiveFormat {
-    TarBzip2,
-    TarGzip,
     TarZstd { config: ZstdConfig },
     TarLz4,
-    Tar,
 }
 
 impl ArchiveFormat {
     /// Get the file extension for the ArchiveFormat
     pub fn extension(&self) -> &str {
         match self {
-            ArchiveFormat::TarBzip2 => TAR_BZIP2_EXTENSION,
-            ArchiveFormat::TarGzip => TAR_GZIP_EXTENSION,
             ArchiveFormat::TarZstd { .. } => TAR_ZSTD_EXTENSION,
             ArchiveFormat::TarLz4 => TAR_LZ4_EXTENSION,
-            ArchiveFormat::Tar => TAR_EXTENSION,
         }
     }
 
@@ -58,13 +45,10 @@ impl TryFrom<&str> for ArchiveFormat {
 
     fn try_from(extension: &str) -> Result<Self, Self::Error> {
         match extension {
-            TAR_BZIP2_EXTENSION => Ok(ArchiveFormat::TarBzip2),
-            TAR_GZIP_EXTENSION => Ok(ArchiveFormat::TarGzip),
             TAR_ZSTD_EXTENSION => Ok(ArchiveFormat::TarZstd {
                 config: ZstdConfig::default(),
             }),
             TAR_LZ4_EXTENSION => Ok(ArchiveFormat::TarLz4),
-            TAR_EXTENSION => Ok(ArchiveFormat::Tar),
             _ => Err(ParseError::InvalidExtension(extension.to_string())),
         }
     }
@@ -107,8 +91,6 @@ mod tests {
 
     #[test]
     fn test_extension() {
-        assert_eq!(ArchiveFormat::TarBzip2.extension(), TAR_BZIP2_EXTENSION);
-        assert_eq!(ArchiveFormat::TarGzip.extension(), TAR_GZIP_EXTENSION);
         assert_eq!(
             ArchiveFormat::TarZstd {
                 config: ZstdConfig::default(),
@@ -117,19 +99,10 @@ mod tests {
             TAR_ZSTD_EXTENSION
         );
         assert_eq!(ArchiveFormat::TarLz4.extension(), TAR_LZ4_EXTENSION);
-        assert_eq!(ArchiveFormat::Tar.extension(), TAR_EXTENSION);
     }
 
     #[test]
     fn test_try_from() {
-        assert_eq!(
-            ArchiveFormat::try_from(TAR_BZIP2_EXTENSION),
-            Ok(ArchiveFormat::TarBzip2)
-        );
-        assert_eq!(
-            ArchiveFormat::try_from(TAR_GZIP_EXTENSION),
-            Ok(ArchiveFormat::TarGzip)
-        );
         assert_eq!(
             ArchiveFormat::try_from(TAR_ZSTD_EXTENSION),
             Ok(ArchiveFormat::TarZstd {
@@ -141,10 +114,6 @@ mod tests {
             Ok(ArchiveFormat::TarLz4)
         );
         assert_eq!(
-            ArchiveFormat::try_from(TAR_EXTENSION),
-            Ok(ArchiveFormat::Tar)
-        );
-        assert_eq!(
             ArchiveFormat::try_from(INVALID_EXTENSION),
             Err(ParseError::InvalidExtension(INVALID_EXTENSION.to_string()))
         );
@@ -152,14 +121,6 @@ mod tests {
 
     #[test]
     fn test_from_str() {
-        assert_eq!(
-            ArchiveFormat::from_str(TAR_BZIP2_EXTENSION),
-            Ok(ArchiveFormat::TarBzip2)
-        );
-        assert_eq!(
-            ArchiveFormat::from_str(TAR_GZIP_EXTENSION),
-            Ok(ArchiveFormat::TarGzip)
-        );
         assert_eq!(
             ArchiveFormat::from_str(TAR_ZSTD_EXTENSION),
             Ok(ArchiveFormat::TarZstd {
@@ -169,10 +130,6 @@ mod tests {
         assert_eq!(
             ArchiveFormat::from_str(TAR_LZ4_EXTENSION),
             Ok(ArchiveFormat::TarLz4)
-        );
-        assert_eq!(
-            ArchiveFormat::from_str(TAR_EXTENSION),
-            Ok(ArchiveFormat::Tar)
         );
         assert_eq!(
             ArchiveFormat::from_str(INVALID_EXTENSION),